Lightning level 1 revised
I made some small ajustments to the lightning set up, the level 1 directional light should be at a value of 0.2 to make it a bit more brighter. Also i adjusted and deleted some meshes that were wrong on the level 1, it does not include floors or other bigger things.

As well as i created materials inside the engine to make the doors and panels that the player cannot walk shine red.